sh-elf.eclass: Add dev-embedded/sh-elf-binutils to BDEPEND
An assembler, which GNU binutils offers one, is needed whenever source
files need to be assembled, which is virtually always the case for
packages being installed to the system globally, as stopping the package
build proceess at assembler code files is rarely useful.  Thus, it is
reasonable to assume that packages requiring the sh-elf toolchain have
hard dependency on not only GCC but binutils as well.

A C compiler, such as GCC, can be used individually, as the compiler's
main job is to translate C code into assembly.  Even if binutils is
absent, 'gcc' is still usable when the '-S' or '-E' option is specified,
and 'cpp', which corresponds to 'gcc -E', is also functional.  After
all, GCC is just a driver program that calls various programs to run
through the complete pipeline of turning source files into binaries,
namely 'cc', 'as', and 'ld'; GCC takes care of 'cc', and binutils
provides 'as' and 'ld'.

Thus, adding sh-elf-binutils to RDEPEND in dev-embedded/sh-elf-gcc is
not the best solution to ensure that packages needing an assembler can
have an assembler available for the build process; otherwise, any users
who wish to use GCC without binutils (which is possible as described
above, though this use case is rare) would still have to unnecessarily
install binutils.

This problem is kind of unique to cross-compiling toolchains not
installed or integrated with the sys-devel/crossdev framework.  binutils
and GCC are both implicit system dependencies declared in the Gentoo
base profile, so they typically do not appear in any dependency
definitions in ebuilds or eclasses since packages can assume their
existence on the system.

Signed-off-by: Yuan Liao <liaoyuan@gmail.com>