dev-embedded/fxsdk: Fix redundant elog when optfeature prints nothing
The original intention of calling 'elog' before 'optfeature_header' was
to split the lengthy header into two lines, but this obviously does not
work as expected when all packages listed in the optfeature group are
installed.  The only straightforward solution seems to be to shorten the
header so it fits into a single line, allowing the 'elog' call to be
eliminated.

Signed-off-by: Yuan Liao <liaoyuan@gmail.com>