dev-embedded/libp7: Avoid redundant udev_reload upon reinstallation
Prior to this commit, udev_reload would be called twice in a row when
the package is being reinstalled or upgraded, which is unnecessary.  In
such a situation, a single udev_reload call in pkg_postinst() would
suffice since the new ebuild's pkg_postinst() is called after the old
ebuild's pkg_postrm().

Signed-off-by: Yuan Liao <liaoyuan@gmail.com>