dev-embedded/gint: Explicitly declare dev-util/cmake in BDEPEND
Building gint involves invoking fxSDK, which in turn invokes CMake.
Because dev-util/cmake is not in RDEPEND of dev-embedded/fxsdk, there
would be no guarantee that CMake exists in the build environment if it
was not explicitly declared in BDEPEND of dev-embedded/gint.

dev-util/cmake should not be declared in RDEPEND of dev-embedded/fxsdk
because in practice, CMake is not a strict runtime dependency of fxSDK.
fxSDK supports (poorly, though) Makefile-based projects, which can be
built without CMake.  If a user only works with Makefile-based projects,
then CMake is not needed at all.

Signed-off-by: Yuan Liao <liaoyuan@gmail.com>