sys-kernel/asahi-edge-kernel: Move Rust toolchain check to pkg_setup
The Rust toolchain required to build the kernel's Rust components may be
installed after this kernel package's pkg_pretend, so checking them in
pkg_pretend is too early.  In a normal package manager execution, the
earliest time when the required Rust toolchain components (as declared
in BDEPEND) should be available is this package's pkg_setup phase.

Signed-off-by: Yuan Liao <liaoyuan@gmail.com>