Revert "dev-embedded/gint: Restrict GCC version to <12"
This reverts commit 0fd684d13c8ffaeb7a46fd103758b6db2d77b568.

sh-elf-gcc 12 per se is not to blame for gint issues; it happens that
native GCC 12 may miscompile sh-elf-gcc 12 (see the linked GCC upstream
bug for more details).  This means that a copy of sh-elf-gcc 12 built by
native GCC 11 probably works with gint, thus the GCC version restriction
may be dropped.

Bug: https://gcc.gnu.org/bugzilla/show_bug.cgi?id=106609
Signed-off-by: Yuan Liao <liaoyuan@gmail.com>