gnome-extra/systemd-presets-gnome: Improve comment for gcr-ssh-agent
Strictly speaking, GNOME Keyring itself is not an ssh-agent, and the
ssh-agent in GNOME Keyring is being sunset [1].  What
gcr-ssh-agent.socket enables is the ssh-agent wrapper in GCR, which is a
different software package from GNOME Keyring.  gcr-ssh-agent was
created as part of the effort to move GNOME Keyring's ssh-agent
functionality out of GNOME Keyring.  A better description of
gcr-ssh-agent is "an ssh-agent that _integrates_ with GNOME Keyring."

A revision bump is not being done because this change has no functional
impact since it only modifies a comment in a systemd preset file.

[1]: https://gitlab.gnome.org/GNOME/gnome-keyring/-/commit/25c5a1982467802fa12c6852b03c57924553ba73

Signed-off-by: Yuan Liao <liaoyuan@gmail.com>